+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 ## 0.1.10 (unreleased)
 
-- Add `pnp_device_import` resource
+- Add `pnp_import_devices` resource
 - Add `fabric_device` resource and data source, this resource now only works with Catalyst Center version 2.3.7.5+ `/sda/fabricDevices`
 - Add `fabric_l3_handoff_ip_transit` resource and data source
 - Add `transitPeerNetworkId` as `id` to `transit_peer_network` resource